backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
boto3>=1.37.0
pydantic>=2.10.0
pydantic-settings>=2.8.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
python-multipart>=0.0.20
